Further iPadOS layout fixes
